Rename _constants.py to _common.py and hoist deferred import
_constants.py contained functions and data structures beyond just
constants; _common.py better reflects its role as shared utilities.
Also moved the deferred _PLAY_HEADING_PARAGRAPH_RE import in
_headings._split_play_heading_paragraph to a top-level import for
consistency — no circular dependency to break.
